+++ libftprintf/printfnumbers.c
@@ -6,7 +6,7 @@
 /*   By: msharpe <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/16 23:17:05 by msharpe           #+#    #+#             */
-/*   Updated: 2018/01/23 20:57:25 by msharpe          ###   ########.fr       */
+/*   Updated: 2018/01/24 22:22:34 by msharpe          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void			ft_per_doi(va_list *list, t_inputinfo *info, t_passinfo *pass)
 {
-	ft_putnbr(va_arg(*list, int), info, pass);
+	ft_putnbr(((long)va_arg(*list, int)), info, pass);
 }
 
 void			ft_per_dup(va_list *list, t_inputinfo *info, t_passinfo *pass)
